refactor: improve conditional evaluation test cleanup and organization
Move conditional evaluation request beans to dedicated file for better organization. Add afterAll hook to cleanup process instances instead of inline cleanup calls. Remove unnecessary if statement from schema verification test and expect array length to be greater than 0 for process instances.
